refactor: move service to subfolder, update namespace
