+++ EmailSender.Backend/EmailSender.Backend.EmailService/Requests/GetSentHistoryRequestHandler.cs
@@ -6,11 +6,11 @@ namespace EmailSender.Backend.EmailService.Requests
     using System.Threading.Tasks;
     using Microsoft.EntityFrameworkCore;
     using Models;
-    using Services;
     using Database;
     using Responses;
     using Shared.Resources;
     using Shared.Exceptions;
+    using Services.SenderService;
 
     public class GetSentHistoryRequestHandler : TemplateHandler<GetSentHistoryRequest, GetSentHistoryResponse>
     {
