+++ EmailSender.Backend/EmailSender.Backend.EmailService/Requests/VerifyEmailRequestHandler.cs
@@ -3,10 +3,10 @@ namespace EmailSender.Backend.EmailService.Requests
     using System;
     using System.Threading;
     using System.Threading.Tasks;
-    using Services;
     using Responses;
     using Shared.Resources;
     using Shared.Exceptions;
+    using Services.SenderService;
 
     public class VerifyEmailRequestHandler : TemplateHandler<VerifyEmailRequest, VerifyEmailResponse>
     {
