+++ EmailSender.WebApi/Configuration/Dependencies.cs
@@ -14,9 +14,11 @@
     using Backend.Shared.Models;
     using Backend.Shared.Behaviours;
     using Backend.Database.Initializer;
-    using Backend.EmailService.Services;
     using Backend.Shared.Services.LoggerService;
     using Backend.Shared.Services.DateTimeService;
+    using Backend.EmailService.Services.UserService;
+    using Backend.EmailService.Services.SenderService;
+    using Backend.EmailService.Services.BillingService;
     using MediatR;
     using DnsClient;
     using MailKit.Net.Smtp;
@@ -32,7 +34,7 @@
             if (environment != null)
                 SetupRetryPolicyWithPolly(services, configuration, environment);
         }
-    
+
         public static void CommonServices(IServiceCollection services, IConfiguration configuration)
         {
             SetupLogger(services);
@@ -40,10 +42,10 @@
             SetupValidators(services);
             SetupMediatR(services);
         }
-    
+
         private static void SetupLogger(IServiceCollection services) 
             => services.AddSingleton<ILoggerService, LoggerService>();
-    
+
         private static void SetupDatabase(IServiceCollection services, IConfiguration configuration) 
         {
             const int maxRetryCount = 10;
@@ -55,20 +57,22 @@
                     => addOptions.EnableRetryOnFailure(maxRetryCount, maxRetryDelay, null));
             });
         }
-    
+
         private static void SetupServices(IServiceCollection services) 
         {
             services.AddHttpContextAccessor();
-    
+
             services.AddScoped<HttpClient>();
             services.AddScoped<ISmtpClient, SmtpClient>();
             services.AddScoped<ILookupClient, LookupClient>();
             services.AddScoped<IDateTimeService, DateTimeService>();
             services.AddScoped<ISenderService, SenderService>();
+            services.AddScoped<IBillingService, BillingService>();
+            services.AddScoped<IUserService, UserService>();
             services.AddScoped<ISmtpClientService, SmtpClientService>();
             services.AddScoped<IDbInitializer, DbInitializer>();
         }
-    
+
         private static void SetupValidators(IServiceCollection services)
             => services.AddValidatorsFromAssemblyContaining<TemplateHandler<IRequest, Unit>>();
